# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Planteo",
        "description": "API pour gérer les utilisateurs, les plantes, et les plantes favorites",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get all plants",
                "operationId": "6e2a3a91967b1e364c87b755cd5152f6",
                "responses": {
                    "200": {
                        "description": "List of plants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Plant"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Create a new plant",
                "operationId": "d005e00370ab9bd321d50ba5dde92042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Plant"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante créée avec succès"
                                        },
                                        "plant": {
                                            "$ref": "#/components/schemas/Plant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/plants/{id}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get a plant by ID",
                "operationId": "5f7b2efcf3ae43a0485b376c550bcdbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Plant"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plant not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Update a plant by ID",
                "operationId": "f433cfe28b4143acc3a1a4a77a7ab646",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Plant"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plant updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La plante a bien été mise à jour."
                                        },
                                        "plant": {
                                            "$ref": "#/components/schemas/Plant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plant not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Delete a plant by ID",
                "operationId": "d1d8619867ca56fe6fa122b59361b856",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La plante a été supprimée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plant not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/plantsbyname": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Search plants by name",
                "operationId": "13f83478afef512775a180fd60f35ec2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Name of the plant",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of plants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Plant"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/plantsbyperiod/{month}/{periodType}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get plants by period type and month",
                "operationId": "1fd1728dbef10ccbe9a5817516500038",
                "parameters": [
                    {
                        "name": "month",
                        "in": "path",
                        "description": "Month as a number (01-12)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "periodType",
                        "in": "path",
                        "description": "Type of period (sowing, planting, harvest)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of plants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Plant"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid period type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid period type"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Plant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Haricots verts"
                    },
                    "description": {
                        "type": "string",
                        "example": "Les haricots verts sont des légumes savoureux et croquants, parfaits pour des salades ou des plats cuisinés."
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "type": {
                        "type": "string",
                        "example": "Légume vert"
                    },
                    "sowing_period": {
                        "type": "string",
                        "example": "03, 04"
                    },
                    "planting_period": {
                        "type": "string",
                        "example": "05, 06"
                    },
                    "harvest_period": {
                        "type": "string",
                        "example": "07, 08, 09"
                    },
                    "soil": {
                        "type": "string",
                        "example": "Sol léger"
                    },
                    "watering": {
                        "type": "string",
                        "example": "Arrosage important pendant la floraison"
                    },
                    "exposure": {
                        "type": "string",
                        "example": "Plein soleil"
                    },
                    "maintenance": {
                        "type": "string",
                        "example": "Biner régulièrement, pailler pour maintenir l'humidité, et butter les plants lorsqu'ils atteignent 10-15 cm"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}